Add files via upload
This script will download, compile with the necessary dependencies and install the XFCE Docklike Plugin Taskbar. This has been tested on Peppermint which is now based on Debian, which I suppose would also work with Debian 11 with the XFCE desktop environment. I have not tested with any Ubuntu based XFCE distros.